Feat: merge navbar & i18n
